+++ mimesis/__init__.py
@@ -1,6 +1,6 @@
 """Mimesis: Fake Data Generator.
 
-Copyright (c) 2016 - 2024 Isaak Uchakaev (Likid Geimfari).
+Copyright (c) 2016 - Present Isaak Uchakaev (Likid Geimfari).
 Website: https://mimesis.name
 Email: <hey@isaak.dev>
 Repository: https://github.com/lk-geimfari/mimesis
@@ -150,7 +150,7 @@ __all__ = [
     "__license__",
 ]
 
-__version__ = "18.0.0"
+__version__ = "19.0.0"
 __title__ = "mimesis"
 __description__ = "Mimesis: Fake Data Generator."
 __url__ = "https://github.com/lk-geimfari/mimesis"