+++ mimesis/builtins/__init__.py
@@ -7,14 +7,12 @@ from .nl import NetherlandsSpecProvider
 from .pl import PolandSpecProvider
 from .pt_br import BrazilSpecProvider
 from .ru import RussiaSpecProvider
-from .uk import UkraineSpecProvider
 
 __all__ = [
     "USASpecProvider",
     "RussiaSpecProvider",
     "BrazilSpecProvider",
     "NetherlandsSpecProvider",
-    "UkraineSpecProvider",
     "PolandSpecProvider",
     "DenmarkSpecProvider",
     "ItalySpecProvider",